==> 2_Investigation of covariates.Rmd <==
---
title: "Selection of covariates"
output:
  github_document:
    toc: true
editor_options:
  chunk_output_type: console
---

# Load in packages and path to folder
```{r echo=T, message=F, error=F, warning=F}
library(dplyr)
library(ggplot2)
library(readr)
library(ggpubr)
library(caret)
library(phyloseq)
library(vegan)
library(lme4)
library(lmerTest)
library(nlme)
library(RCurl)
library(MuMIn)
path = "C:/Users/FX76TZ/OneDrive - Aalborg Universitet/IBD project/V1/data/"
```

# Investigate the proportion of reads removed as human

Read in meta data: 
```{r}
meta <- read.delim(paste0(path, "metadata/", "preped_metadata.txt"))
```

Read in the data:
```{r echo=T, message=F, error=F, warning=F}
#Percentage of removed reads due to human
removed_human <- read_delim(file=paste0(path, "../analysis/", "summary_stats_of_QC.txt"), delim="\t")
removed_human <- removed_human %>% dplyr::select(J_ID, percent_removed_as_human)
removed_human <-  removed_human %>% mutate(percent_removed_as_human = as.numeric(percent_removed_as_human))
summary(removed_human$percent_removed_as_human)

#Unmapped MetaPhlan
bug_relab <- read.csv(paste0(path, "humann3_processed_output/Joined_tables/", "Joined_metaphlan_bugs_list_relab.tsv"), sep = "\t", strip.white = T, stringsAsFactors = F, row.names = 1) 
colnames(bug_relab) <- substr(colnames(bug_relab), start = 1, stop = 6)

relab_unknown <- bug_relab[1,]
relab_unknown <- t(relab_unknown)
relab_unknown <- as.data.frame(relab_unknown)
relab_unknown$J_ID <- rownames(relab_unknown)
```

```{r}
#Merge 
missing_data <- merge(removed_human, relab_unknown, by.x="J_ID", by.y="J_ID")
missing_data <- missing_data %>% rename(metaphlan_unknown=UNKNOWN)
head(missing_data)

#Save missing
write.table(missing_data, file=paste0(path, "metadata/", "missing_data.txt"), col.names = T, row.names = F, quote = F, sep = '\t')

#Plots: Unmapped reads vs. the proportion of reads removed as human
ggplot(missing_data)+
  geom_point(aes(x=percent_removed_as_human*100, y=metaphlan_unknown), size=1)+theme_classic()+
  xlab("Reads removed as human (%)")+ylab("MetaPhlAn: Unmapped reads (%)")+
  geom_vline(xintercept = 85, linetype=2, color="blue")

sum(missing_data$percent_removed_as_human>0.85) #6 datapoint have more that 85% of reads removed as human reads

#Remove samples with more than 85% removed as human:
meta <- merge(meta, missing_data, by.x="J_ID", by.y="J_ID")
meta <- meta %>% filter(percent_removed_as_human < 0.85)
dim(meta) #202 samples left

#Save the new meta data:
write.table(meta, file=paste0(path, "metadata/", "preped_metadata_filtered.txt"), col.names = T, row.names = F, quote = F, sep = '\t')

#Read in data and subset the phyloseq objects based on the filtered meta data:
ps_count <- readRDS(paste0(path, "humann3_processed_output/Phyloseq/","ps_count_filtered.rds"))
ps_relab <- readRDS(paste0(path, "humann3_processed_output/Phyloseq/","ps_relab_filtered.rds")) 

ps_count_sub <-subset_samples(ps_count, J_ID %in% meta$J_ID)
ps_relab_sub <-subset_samples(ps_relab, J_ID %in% meta$J_ID) 

saveRDS(ps_count_sub, paste0(path, "humann3_processed_output/Phyloseq/","ps_count_filtered2.rds"))
saveRDS(ps_relab_sub, paste0(path, "humann3_processed_output/Phyloseq/","ps_relab_filtered2.rds"))
```

Look into the samples left:
```{r}
meta_flare <- meta %>% filter(f_cal_current > 250)
meta_flare_c <- meta_flare %>% filter(age_gr==1)
length(unique(meta_flare_c$person_ID))
meta_flare_a <- meta_flare %>% filter(age_gr==2)
length(unique(meta_flare_a$person_ID))

#Split samples by timepoint:
t0 <- meta %>% filter(time_point==0)
dim(t0)
t1 <- meta %>% filter(time_point==1)
dim(t1)
t2 <- meta %>% filter(time_point==2)
dim(t2)
t3 <- meta %>% filter(time_point==3)
dim(t3)
t4 <- meta %>% filter(time_point==4)
dim(t4)

#How many female/male (0/1) in each?
dim(t0 %>% filter(sex==0))
dim(t0 %>% filter(sex==1))
dim(t1 %>% filter(sex==0))
dim(t1 %>% filter(sex==1))
dim(t2 %>% filter(sex==0))
dim(t2 %>% filter(sex==1))
dim(t3 %>% filter(sex==0))
dim(t3 %>% filter(sex==1))
dim(t4 %>% filter(sex==0))
dim(t4 %>% filter(sex==1))

#How many children/adults (1/2) in each?
dim(t0 %>% filter(age_gr==1))
dim(t0 %>% filter(age_gr==2))
dim(t1 %>% filter(age_gr==1))
dim(t1 %>% filter(age_gr==2))
dim(t2 %>% filter(age_gr==1))
dim(t2 %>% filter(age_gr==2))
dim(t3 %>% filter(age_gr==1))
dim(t3 %>% filter(age_gr==2))
dim(t4 %>% filter(age_gr==1))
dim(t4 %>% filter(age_gr==2))

#How many with remission/mild/moderate/severe disease?
dim(t0 %>% filter(score_num==1))
dim(t0 %>% filter(score_num==2))
dim(t0 %>% filter(score_num==3))
dim(t0 %>% filter(score_num==4))
dim(t1 %>% filter(score_num==1))
dim(t1 %>% filter(score_num==2))
dim(t1 %>% filter(score_num==3))
dim(t1 %>% filter(score_num==4))
dim(t2 %>% filter(score_num==1))
dim(t2 %>% filter(score_num==2))
dim(t2 %>% filter(score_num==3))
dim(t2 %>% filter(score_num==4))
dim(t3 %>% filter(score_num==1))
dim(t3 %>% filter(score_num==2))
dim(t3 %>% filter(score_num==3))
dim(t3 %>% filter(score_num==4))
dim(t4 %>% filter(score_num==1))
dim(t4 %>% filter(score_num==2))
dim(t4 %>% filter(score_num==3))
dim(t4 %>% filter(score_num==4))

#How many children with different clinical disease scores?
data_child <- meta %>% filter(age_gr==1)
length(unique(data_child$person_ID))
dim(data_child %>% filter(score_num==1))
dim(data_child %>% filter(score_num==2))
dim(data_child %>% filter(score_num==3))
dim(data_child %>% filter(score_num==4))

#How many adults?
data_adult <- meta %>% filter(age_gr==2)
length(unique(data_adult$person_ID))
dim(data_adult %>% filter(score_num==1))
dim(data_adult %>% filter(score_num==2))
dim(data_adult %>% filter(score_num==3))
dim(data_adult %>% filter(score_num==4))

#Summary stats of f-cal
summary(data_child$f_cal_current)
summary(data_adult$f_cal_current)
```


# Investigate predicters: Correlation and distribution

Investigate VIF and correlation coefficients
```{r}
#Model of the 3 main covariates: 
preds <- meta %>% dplyr::select(f_cal_current, age_gr, study_gr,sex)
model <- lm(f_cal_current~., data=preds)
car::vif(model)

#All predictors and disease score
preds <- meta %>% dplyr::select(f_cal_current, age_gr, study_gr,sex, study_gr,sex, asa, aza, pred, l_asa, l_pred, bio)
model <- lm(f_cal_current~., data=preds)
car::vif(model)
#Largest VIF = 1.65 - not high enough to remove predictors

#Investigate correlation coefficients:
preds <- meta %>% dplyr::select(f_cal_current, age_gr, study_gr,sex, study_gr,sex, asa, aza, pred, l_asa, l_pred, bio)
#Spearman correlation
df2 <- cor(preds, method = "spearman", use = "pairwise.complete.obs")
df2
#Pearson correlation:
df2 <- cor(preds, method = "pearson", use = "pairwise.complete.obs")
df2


#How similar are clinical and paraclinical remission?
meta$score_binary_compare <- as.factor(ifelse(meta$score_binary == "remission", 0, 1))

sum(as.numeric(meta$fcal_binary==meta$score_binary_compare), na.rm=T)/(dim(meta)[1]-sum(is.na(as.numeric(meta$fcal_binary==meta$score_binary_compare))))

#How similar are the discrete numeric disease score and the measure of f_cal?
preds <- meta %>% dplyr::select(f_cal_current, score_num)
df2 <- cor(preds, method = "spearman", use = "pairwise.complete.obs")
df2

```


Are disease score and numeric predictors normally distributed?
```{r warning=FALSE, message=F}
#Check faecal calprotectin:
ggplot(meta)+geom_histogram(aes(x=f_cal_current), fill="lightblue", color="black")+theme_classic()

#Check discrete numeric disease score:
ggplot(meta)+geom_histogram(aes(x=score_num), fill="lightblue", color="black")+theme_classic()
```

The two disease scores are not normally distributed, and this must be considered in further analyses.
